requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
aspose-words==23.8.0
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
chroma-hnswlib
chromadb==0.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.5.14
Django==4.2.4
docker-pycreds==0.4.0
exceptiongroup==1.1.3
fastapi==0.99.1
ffmpy==0.3.1
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gradio==3.40.1
gradio_client==0.4.0
greenlet==2.0.2
gspread==5.10.0
h11==0.14.0
html2text==2020.1.16
httpcore==0.17.3
httplib2==0.22.0
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-resources==6.0.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
langchain==0.0.268
langsmith==0.0.25
linkify-it-py==2.0.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numexpr==2.8.5
numpy==1.25.2
oauth2client==4.1.3
oauthlib==3.2.2
onnxruntime==1.15.1
openai==0.27.8
orjson==3.9.5
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
Pillow==10.0.0
posthog==3.0.2
protobuf==4.24.1
psutil==5.9.5
pulsar-client==3.2.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.12
pydantic_core==2.6.1
pydub==0.25.1
pyparsing==3.0.9
PyPika==0.48.9
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.9.2
rsa==4.9
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.11.2
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.29.2
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
toolz==0.12.0
tqdm==4.66.1
transformers==4.31.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==1.26.16
uvicorn==0.23.2
wandb==0.15.8
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
gunicorn
beautifulsoup4
html2text